Move terminal helpers to standardization
