+++ datasets/openthoughts_agent_sft/raw_to_atif.py
@@ -4,12 +4,7 @@ import json
 import sys
 from typing import Any
 
-from scripts.raw_to_atif_common import (
-    dataset_name_from_script,
-    split_terminal_task_description_prompt,
-    structure_terminal_completion_step,
-    trajectory_from_record,
-)
+from scripts.raw_to_atif_common import dataset_name_from_script, trajectory_from_record
 
 
 def main(script_file: str) -> None:
@@ -21,9 +16,6 @@ def main(script_file: str) -> None:
         if not isinstance(record, dict):
             continue
         trajectory = trajectory_from_record(record, index, dataset_name)
-        split_terminal_task_description_prompt(trajectory)
-        for step in trajectory.steps:
-            structure_terminal_completion_step(step)
         print(trajectory.model_dump_json(exclude_none=True))
