+++ scripts/atif_to_std_common.py
@@ -7,7 +7,13 @@ import re
 import sys
 from typing import Any
 
-from schema.atif import ATIFTrajectory, ToolCall, normalize_atif_trajectory
+from schema.atif import (
+    ATIFTrajectory,
+    Step,
+    ToolCall,
+    content_to_text,
+    normalize_atif_trajectory,
+)
 
 SYSTEM_PROMPT_PREFIXES = (
     "You are web shopping.\n",
@@ -19,6 +25,7 @@ RESPONSE_FORMAT_PROMPT = re.compile(
     r"\nYour response should use the following format:.*\Z",
     re.DOTALL,
 )
+TERMINAL_TASK_DESCRIPTION_MARKER = "\n\nTask Description:\n"
 
 
 def _string_arg(arguments: dict[str, Any], *names: str) -> str:
@@ -101,6 +108,72 @@ def normalize_prompt_boilerplate(trajectory: ATIFTrajectory) -> None:
             step.source = "system"
             continue
         step.message = RESPONSE_FORMAT_PROMPT.sub("", step.message).strip()
+
+
+def renumber_steps(steps: list[Step]) -> list[Step]:
+    for index, step in enumerate(steps, start=1):
+        step.step_id = index
+    return steps
+
+
+def split_terminal_task_description_prompt(trajectory: ATIFTrajectory) -> bool:
+    if not trajectory.steps:
+        return False
+    first_step = trajectory.steps[0]
+    if first_step.source != "user" or not isinstance(first_step.message, str):
+        return False
+    if TERMINAL_TASK_DESCRIPTION_MARKER not in first_step.message:
+        return False
+    system_prompt, task_prompt = first_step.message.split(
+        TERMINAL_TASK_DESCRIPTION_MARKER, 1
+    )
+    first_step.source = "system"
+    first_step.message = system_prompt.strip()
+    trajectory.steps.insert(
+        1,
+        Step(
+            step_id=2,
+            source="user",
+            message=f"Task Description:\n{task_prompt.strip()}",
+        ),
+    )
+    trajectory.steps = renumber_steps(trajectory.steps)
+    return True
+
+
+def structure_terminal_completion_step(step: Step) -> bool:
+    if step.source != "agent" or step.tool_calls or step.observation is not None:
+        return False
+    text = content_to_text(step.message).strip()
+    if not text.startswith("{"):
+        return False
+    try:
+        payload = json.loads(text)
+    except json.JSONDecodeError:
+        return False
+    if not isinstance(payload, dict) or payload.get("task_complete") is not True:
+        return False
+    commands = payload.get("commands")
+    if not isinstance(commands, list) or any(
+        isinstance(command, dict) and str(command.get("keystrokes") or "").strip()
+        for command in commands
+    ):
+        return False
+    message_parts = [
+        str(payload.get("analysis") or "").strip(),
+        str(payload.get("plan") or "").strip(),
+    ]
+    message = "\n\n".join(part for part in message_parts if part)
+    step.message = message
+    step.tool_calls = [
+        ToolCall(
+            tool_call_id="call_1",
+            function_name="finish",
+            arguments={"message": message, "task_completed": True},
+            extra={"raw_format": "terminal_json"},
+        )
+    ]
+    return True
 
 
 def standardize_tools(trajectory: ATIFTrajectory) -> ATIFTrajectory:
