+++ scripts/raw_to_atif_common.py
@@ -74,7 +74,6 @@ PROMPT_FIELDS = (
     "sop",
 )
 RESPONSE_FIELDS = ("response", "answer", "output", "solution")
-TERMINAL_TASK_DESCRIPTION_MARKER = "\n\nTask Description:\n"
 
 
 def dataset_name_from_script(script_file: str) -> str:
@@ -740,65 +739,6 @@ def renumber_steps(steps: list[Step]) -> list[Step]:
     for index, step in enumerate(steps, start=1):
         step.step_id = index
     return steps
-
-
-def split_terminal_task_description_prompt(trajectory: ATIFTrajectory) -> bool:
-    if not trajectory.steps:
-        return False
-    first_step = trajectory.steps[0]
-    if first_step.source != "user" or not isinstance(first_step.message, str):
-        return False
-    if TERMINAL_TASK_DESCRIPTION_MARKER not in first_step.message:
-        return False
-    system_prompt, task_prompt = first_step.message.split(
-        TERMINAL_TASK_DESCRIPTION_MARKER, 1
-    )
-    first_step.source = "system"
-    first_step.message = system_prompt.strip()
-    trajectory.steps.insert(
-        1,
-        Step(
-            step_id=2,
-            source="user",
-            message=f"Task Description:\n{task_prompt.strip()}",
-        ),
-    )
-    trajectory.steps = renumber_steps(trajectory.steps)
-    return True
-
-
-def structure_terminal_completion_step(step: Step) -> bool:
-    if step.source != "agent" or step.tool_calls or step.observation is not None:
-        return False
-    text = text_from_content(step.message).strip()
-    if not text.startswith("{"):
-        return False
-    try:
-        payload = json.loads(text)
-    except json.JSONDecodeError:
-        return False
-    if not isinstance(payload, dict) or payload.get("task_complete") is not True:
-        return False
-    commands = payload.get("commands")
-    if not isinstance(commands, list) or any(
-        isinstance(command, dict) and str(command.get("keystrokes") or "").strip()
-        for command in commands
-    ):
-        return False
-    message_parts = [
-        str(payload.get("analysis") or "").strip(),
-        str(payload.get("plan") or "").strip(),
-    ]
-    step.message = "\n\n".join(part for part in message_parts if part)
-    step.tool_calls = [
-        ToolCall(
-            tool_call_id="call_1",
-            function_name="finish",
-            arguments={"message": step.message, "task_completed": True},
-            extra={"raw_format": "terminal_json"},
-        )
-    ]
-    return True
 
 
 def screenagent_trajectories(items: list[Any], dataset_name: str) -> Iterable[ATIFTrajectory]:
